go mod

+++ engine/config.go
@@ -8,5 +8,5 @@ type Config struct {
 	EnableSeeding     bool
 	IncomingPort      int
 	DoneCmd           string
-	SeedRatio		  float64
+	SeedRatio		  float32
 }
