+++ engine/torrent.go
@@ -21,7 +21,7 @@ type Torrent struct {
 	DoneCmdCalled bool
 	Percent       float32
 	DownloadRate  float32
-	SeedRatio	  float64
+	SeedRatio	  float32
 	Stats         torrent.TorrentStats
 	t             *torrent.Torrent
 	updatedAt     time.Time
@@ -45,7 +45,7 @@ func (torrent *Torrent) Update(t *torrent.Torrent) {
 	if torrent.Loaded {
 		torrent.updateLoaded(t)
 		torrent.Stats = t.Stats()
-		torrent.SeedRatio = float64(torrent.Stats.BytesWritten.Int64()) / float64(torrent.Stats.BytesReadData.Int64())
+		torrent.SeedRatio = float32(torrent.Stats.BytesWritten.Int64()) / float32(torrent.Stats.BytesReadData.Int64())
 	}
 	torrent.t = t
 }
